+++ docs/whats_new.rst
@@ -74,6 +74,10 @@ Enhancements
 Deprecations
 ~~~~~~~~~~~~
 
+- The ``regionmask.defined_regions._ar6_pre_revisions`` regions are deprecated. The
+  ``regionmask.defined_regions.ar6`` regions should be used instead
+  (:issue:`314`, :pull:`320`).
+
 New regions
 ~~~~~~~~~~~
 
@@ -88,6 +92,8 @@ Bug Fixes
 - The name of lon and lat coordinates when passed as single elements is now repected when
   creating masks i.e. for ``region.mask(ds.longitude, ds.longitude)`` (:issue:`129`,
   :pull:`294`).
+- Ensure :py:meth:`Regions.plot` uses the current axes (``plt.gca()``) if possible and
+  error if a non-cartopy GeoAxes is passed (:issue:`316`, :pull:`321`).
 
 
 Docs